requirements.txt
alembic==1.4.2
aniso8601==9.0.1
APScheduler==3.6.3
attrs==19.3.0
awscli==1.18.20
bcrypt==3.1.7
botocore==1.15.20
certifi==2019.9.11
cffi==1.15.0
chardet==3.0.4
checksumdir==1.1.7
Click==7.0
colorama==0.4.3
dnspython==1.16.0
docutils==0.15.2
filechunkio==1.8
Flask==1.1.1
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.8
Flask-Executor==0.9.3
Flask-JWT-Extended==3.24.1
flask-marshmallow==0.14.0
Flask-Migrate==2.5.2
flask-restx==0.5.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
fpdf==1.7.2
greenlet==1.1.2
html5lib==1.0.1
idna==2.8
importlib-metadata==0.23
inflect==5.2.0
itsdangerous==1.1.0
Jinja2==2.10.3
jmespath==0.9.5
jsonschema==3.1.1
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.10.0
more-itertools==7.2.0
mtools==1.6.3
pdfkit==0.6.1
phonenumbers==8.12.18
Pillow==9.0.1
pisa==3.0.33
psutil==5.7.0
psycopg2==2.9.3
public==2019.4.13
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
pymongo==3.11.3
PyPDF2==1.26.0
pyrsistent==0.15.5
python-dateutil==2.8.1
python-dotenv==0.10.3
python-editor==1.0.4
python-resize-image==1.1.19
pytz==2019.3
PyYAML==5.3
query-string==2019.4.13
reportlab==3.6.9
requests==2.22.0
rsa==3.4.2
s3transfer==0.3.3
schedule==0.6.0
selenium==3.8.1
six==1.12.0
SQLAlchemy==1.3.19
tzlocal==2.0.0
urllib3==1.25.6
virtualenv==16.7.5
webencodings==0.5.1
Werkzeug==0.16.0
WTForms==2.2.1
xhtml2pdf==0.2.4
zipp==0.6.0
